seq: use guint8 for the numerical ID of queue

+++ src/seq/query.h
@@ -43,13 +43,13 @@ void alsaseq_get_subscription_list(const ALSASeqAddr *addr,
 void alsaseq_get_queue_id_list(guint **entries, gsize *entry_count,
                                GError **error);
 
-void alsaseq_get_queue_info_by_id(guint queue_id, ALSASeqQueueInfo **queue_info,
+void alsaseq_get_queue_info_by_id(guint8 queue_id, ALSASeqQueueInfo **queue_info,
                                   GError **error);
 void alsaseq_get_queue_info_by_name(const gchar *name,
                                     ALSASeqQueueInfo **queue_info,
                                     GError **error);
 
-void alsaseq_get_queue_status(guint queue_id,
+void alsaseq_get_queue_status(guint8 queue_id,
                               ALSASeqQueueStatus *const *queue_status,
                               GError **error);
